test-app/requirements.txt
Flask==3.1.3
requests==2.32.3
python-dotenv==1.2.1
